Add Meson build configuration
The Meson build reproduce exactly the configure-based build except
for the version naming of the GSL library when built as a shared library.

In addition the gslcblas, when used, will be always statically linked
to the gsl library itself.

If the meson option -Dblas=<some-blas> is used an external BLAS library
is used instead of gslcblas.